backend/requirements.txt

Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
django-filter==23.5
factory-boy==3.3.0
Faker==22.5.1
gunicorn==21.2.0
django-throttling==1.1.1
